requirements.txt
scapy
PyYAML
pandas
tabulate
